+++ src/extensions/AssetView.ts
@@ -529,7 +529,7 @@ export class AssetView
                 }
             }
 
-            options.push( null, { name: 'Delete', icon: 'Trash2', className: 'text-destructive',
+            options.push( null, { name: 'Delete', icon: 'Trash2', className: 'destructive',
                 callback: that._requestDeleteItem.bind( that, item ) } );
 
             LX.addClass( that.contentPanel.root, 'pointer-events-none' );