+++ changelog.md
@@ -3,6 +3,8 @@
 ## dev
 
 Fixed TextInput not updating on browser autofill.
+Allow support to disable: Array, DatePicker, Layers, List, Pad, Rate, and Tags components.
+Add `options.mapIcon` to change Map2D opener button icon.
 
 Form Component:
 - Enter key now submits the form as native behaviour.